Latest build
This links the jot_home.php to the Login_screen.php and also copys the formating and buttons to navagate back to the page. However the create and contact us pages are bugy. The contact us page does not exist yet and the create user page is still having issues working with the mysql database/table.